Add HomeView with stats
The HomeView is the main landing page, and includes several general
statistics about the currently stored Dandisets.

I included a TODO in `views/HomeView/Stat.vue`. The original spec called
for 1px dividers between the statistics. I added these, but their
presence disturbs the grid system; it will always overflow the last
statistic to a new line, which disturbs the breakpoints set up for
low resolutions. I opted to remove the dividers in favor of a cleaner UX
on small screens.